requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.72.0
anyio==4.11.0
attrs==25.4.0
Authlib==1.6.5
backoff==2.2.1
beautifulsoup4==4.14.2
browser-use==0.9.5
bubus==1.5.6
cachetools==6.2.1
cdp-use==1.4.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.2
cryptography==46.0.3
Cython==3.2.0
distro==1.9.0
docstring_parser==0.17.0
frozenlist==1.8.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-genai==1.49.0
googleapis-common-protos==1.72.0
groq==0.33.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
inquirerpy==0.3.4
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain-core==1.0.4
langchain-openai==1.0.2
langsmith==0.4.42
markdown-it-py==4.0.0
markdownify==1.2.0
mcp==1.21.0
mdurl==0.1.2
multidict==6.7.0
oauthlib==3.3.1
ollama==0.6.0
openai==1.109.1
orjson==3.11.4
packaging==25.0
pfzy==0.3.4
pillow==12.0.0
portalocker==2.10.1
posthog==6.9.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyobjc-core==12.0
pyobjc-framework-Cocoa==12.0
pyotp==2.9.0
pyparsing==3.2.5
pypdf==6.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
reportlab==4.4.4
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
screeninfo==0.8.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sse-starlette==3.0.3
starlette==0.50.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uuid7==0.1.0
uvicorn==0.38.0
wcwidth==0.2.14
websockets==15.0.1
yarl==1.22.0
zstandard==0.25.0
